+++ src/Event/HotCodeReloaderWorkerStartListener.php
@@ -10,22 +10,51 @@ declare(strict_types=1);
 
 namespace Mezzio\Swoole\Event;
 
-use Mezzio\Swoole\HotCodeReload\Reloader;
+use Mezzio\Swoole\HotCodeReload\FileWatcherInterface;
+use Psr\Log\LoggerInterface;
 
 class HotCodeReloaderWorkerStartListener
 {
-    private Reloader $reloader;
+    /**
+     * A file watcher to monitor changes in files.
+     *
+     * @var FileWatcherInterface
+     */
+    private $fileWatcher;
 
-    public function __construct(Reloader $reloader)
+    /** @var LoggerInterface */
+    private $logger;
+
+    /** @var int */
+    private $interval;
+
+    public function __construct(FileWatcherInterface $fileWatcher, LoggerInterface $logger, int $interval)
     {
-        $this->reloader = $reloader;
+        $this->fileWatcher = $fileWatcher;
+        $this->logger      = $logger;
+        $this->interval    = $interval;
     }
 
     public function __invoke(WorkerStartEvent $event): void
     {
-        $this->reloader->onWorkerStart(
-            $event->getServer(),
-            $event->getWorkerId()
-        );
+        if (0 !== $event->getWorkerId()) {
+            return;
+        }
+
+        $server      = $event->getServer();
+        $fileWatcher = $this->fileWatcher;
+        $logger      = $this->logger;
+
+        $server->tick($this->interval, function () use ($server, $fileWatcher, $logger) {
+            $changedFilePaths = $fileWatcher->readChangedFilePaths();
+            if (! $changedFilePaths) {
+                return;
+            }
+
+            foreach ($changedFilePaths as $path) {
+                $logger->notice('Reloading due to file change: {path}', ['path' => $path]);
+            }
+            $server->reload();
+        });
     }
 }